implement enveloped tagging
